refactor: Rename the project to rscm
