+++ Cargo.toml
@@ -1,10 +1,10 @@
 [package]
-name = "two-layer-model"
+name = "rscm"
 version = "0.1.0"
 edition = "2021"
 
 [lib]
-name = "two_layer_model"
+name = "rscm"
 # "cdylib" is necessary to produce a shared library for Python to import from.
 crate-type = ["cdylib", "rlib"]
