+++ pyproject.toml
@@ -6,12 +6,12 @@ build-backend = 'maturin'
 
 [tool.maturin]
 python-source = "python"
-module-name = "two_layer_model._lib"
+module-name = "rscm._lib"
 bindings = 'pyo3'
 features = ["pyo3/extension-module"]
 
 [project]
-name = "two-layer-model"
+name = "rscm"
 version = "0.1.0"
 description = "Two layer model implemented in Rust"
 readme = "README.md"
