+++ python/rscm/__init__.py
@@ -1,0 +1,7 @@
+"""
+Rust Simple Climate Model (RSCM)
+
+A framework for simple climate models built it Rust.
+"""
+
+from ._lib import TwoLayerComponentBuilder  # noqa
