+++ python/rscm/_lib/__init__.pyi
