+++ python/rscm/_lib/components.pyi
@@ -1,4 +1,4 @@
-from two_layer_model._lib.core import ComponentBuilder
+from rscm._lib.core import ComponentBuilder
 
 class CarbonCycleBuilder(ComponentBuilder): ...
 class CO2ERFBuilder(ComponentBuilder): ...
