+++ python/rscm/components.py
@@ -2,7 +2,7 @@
 RSCM Components defined in Rust
 """
 
-from two_layer_model._lib.components import CarbonCycleBuilder, CO2ERFBuilder
+from rscm._lib.components import CarbonCycleBuilder, CO2ERFBuilder
 
 __all__ = [
     "CO2ERFBuilder",
