--- python/two_layer_model/__init__.py
@@ -1,5 +1,0 @@
-"""
-Two Layer Model
-"""
-
-from ._lib import TwoLayerComponentBuilder  # noqa
